Reduce number of SELECTs when importing status entries.
We don't have to run a SELECT for each status entry we're importing into
the database. Instead, we can check whether there are any status entries
for a given valid-after time, and if not, keep track of imported rows
ourselves.

This reduces the number of SELECTs for importing a consensus by a factor
of 2000 (or whatever number of status entries there are in a consensus).