Remove unused GeoIPDatabaseManager.
We once needed to mantain different GeoIP database versions to resolve
bridge IP addresses to country codes in the sanitizing process.  We gave
up on that, because it was too complex.  That means we don't need the
GeoIP database manager anymore.  If we decide we need it, it'll live on in
Git's history.