Move materialized view update logic from the database to Java.
So far, we used row-level triggers to memorize which dates had new or
updated data and needed to be included in the next materialized view
refresh run.  The row-level triggers wrote these dates to a table which
was read by refresh functions.  After a successful refresh run, the table
was emptied.  There were two problems with this approach:

First, the refresh run is executed in a single long-running transaction.
If someone adds new data to the database while a refresh run is active,
the changed dates won't be included in the next refresh run.

Second, row-level triggers have a negative impact on performance.  It
turns out that removing them reduces import time by roughly 18 %.

The new approach uses Java to memorize which dates need to be updated.