Exclude bridge network statuses from duplicates check
There's simply no way for us to detect whether two (subsequent) statuses
are legitimately equal or not, so we'll have to consider them valid.  The
only thing we can do is warn if referenced descriptors are older than a
given number of hours and ask the operator to investigate manually.  (This
check will be contained in a separate commit.)