Warn if a bridge network status is possibly stale.
In February 2011, Tonga's tor process died, but it's tar cronjob continued
to suppy us with "fresh" bridge descriptor tarballs (containing stale
descriptors).  We didn't detect this for two weeks, because we derive the
bridge network status publication time from the tarball filename, and it
looked recent all the time.

Adding a check whether the most recent descriptor that is contained in a
bridge network status was published more than 1 hour before the status
itself.  This is a sign for the status being stale, even though it's not
100% reliable.  We still should investigate these cases when they occur.

A test with the January and February 2011 tarballs shows that we'd have
learned about Tonga's tor process dying 60-90 minutes after the fact.