Download all server/extra-info descriptors once per day.
The relay descriptor downloader so far only downloads server and
extra-info descriptors that are referenced in other directory documents
and that we are missing.

With this patch we also download /tor/server/all and /tor/extra/all from
the directories once per day to learn about server and extra-info
descriptors that are not referenced anywhere.

This patch further requests the compressed versions of descriptors by
adding ".z" to the URLs in order to save bandwidth.

This patch is a 50% rewrite of the relay descriptor downloader which
became too complex over time.