Download votes when v3-status-votes file is truncated.
In the past, we sometimes had problems rsync'ing the v3-status-votes file
from gabelmoo and were left with a truncated file.  When parsing this file
we successfully parsed the complete votes and failed parsing the last,
truncated vote.  But we however marked this vote as downloaded and made no
further attempts to download it from the directory authorities.

The fix is to only mark a vote as downloaded if it contains a valid
directory footer.