Sanity-check relay descriptors before writing them to disk.
In #5805 we found that we have quite a few files that are either empty or
truncated.  Turns out that the current metrics-db code doesn't allow
writing empty files, but it does allow writing truncated files.  We now
parse all descriptors with metrics-lib and only store valid descriptors to
disk.  Fixes part of #5813.