Remember longer than 2 days that we stored a microdesc.
Our internal descriptor completeness statistics just broke, because we
stopped remembering that we stored a microdesc more than 2 days ago.  For
server descriptors and extra-info descriptors 2 days are sufficient,
because they won't be referenced in a consensus after that.  But
microdescs may be used for a week or more.  In fact, raise period to 40
days to be sure that we downloaded microdescs this month.  This only
affects internal statistics, nothing else.