Consider only newly stored descriptors in stale check.
A few days ago there was an unusual situation with parsing gabelmoo's
cached descriptors and fetching descriptors from the other directory
authorities.  gabelmoo failed to produce a consensus at :00 of the hour
and downloaded the consensus from the other directory authorities after
:05 of the hour.  As a result, gabelmoo's cached consensus was 1 hour
older than the consensus that we downloaded from the other directory
authorities.

Our stale check considers all stored descriptors and determines if there
were any consensuses for which we didn't also store votes or microdesc
consensuses.  In this case the stale check triggered, because we didn't
store any votes or microdesc consensuses with the same valid-after time as
gabelmoo's outdated cached consensus.  This warning was confusing.

The fix is simple: if we're storing a descriptor that we stored before,
don't include it in the stale check.