Store descriptors that metrics-lib cannot parse.
So far have been using metrics-lib to parse relay descriptors prior to
storing them.  But that turned out to be a bad idea.  If the directory
authorities serve these descriptors, and if we can in theory serve them,
too, we should store and serve them.  Leaving in the check, but rather
than skipping the descriptor, storing it anyway.

Implements #19170.