Correctly index files that are moved away and back.
The indexer did not handle a (mostly theoretic) edge case of a file
being moved away and then moved back shortly after. In such a case the
file should not be marked for deletion anymore and it should be
included in the index again. That's what this commit does.

The other minor changes to unit tests are just cosmetic.

Fixes #34030.